refactor(app): extract conductor main gradio client
