+++ apps/conductor-main/pyproject.toml
@@ -1,0 +1,28 @@
+[build-system]
+requires = ["setuptools>=69", "wheel"]
+build-backend = "setuptools.build_meta"
+
+[project]
+name = "conductor-main"
+version = "0.1.0"
+description = "Gradio client for Conductor Core"
+readme = "README.md"
+requires-python = ">=3.10"
+dependencies = [
+    "conductor-core",
+    "gradio==5.30.0",
+    "mido==1.3.3",
+    "plotly==6.0.1",
+]
+
+[project.optional-dependencies]
+dev = ["pytest>=8.3,<9"]
+
+[project.scripts]
+conductor-main = "conductor_main.app:main"
+
+[tool.setuptools]
+package-dir = {"" = "src"}
+
+[tool.setuptools.packages.find]
+where = ["src"]
