+++ pyproject.toml
@@ -3,7 +3,7 @@ requires = ["setuptools>=69", "wheel"]
 build-backend = "setuptools.build_meta"
 
 [project]
-name = "loopgpt"
+name = "conductor"
 version = "0.1.0"
 description = "AI-powered 4-bar MIDI loop generation from natural-language prompts"
 readme = "README.md"
@@ -45,8 +45,17 @@ include = ["src*", "evaluation*"]
 namespaces = true
 
 [tool.pytest.ini_options]
-testpaths = ["tests", "packages/conductor-core/tests"]
-pythonpath = ["packages/conductor-core/src"]
+testpaths = [
+	"tests",
+	"packages/conductor-core/tests",
+	"projects/conductor-eval/tests",
+	"apps/conductor-main/tests",
+]
+pythonpath = [
+	"packages/conductor-core/src",
+	"projects/conductor-eval/src",
+	"apps/conductor-main/src",
+]
 python_files = ["test_*.py"]
 addopts = "-ra"
 markers = [
